requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
slack-sdk>=3.23.0
pydantic>=2.5.0
cryptography>=41.0.0
httpx>=0.25.0
python-multipart>=0.0.6
google-generativeai>=0.3.0

